pyproject.toml
[project]
name = "story-mcp-hub"
version = "0.1.0"
description = "Story Protocol MCP Hub"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # MCP dependencies
    "mcp[cli]>=1.2.1",
    "python-dotenv>=1.0.1",
    
    # Web3 and Story Protocol SDK
    "web3",
    "story-protocol-python-sdk",
    
    # HTTP and API dependencies
    "httpx>=0.28.1",
    "requests>=2.32.3",
    "aiohttp>=3.11.12",
    "uvicorn>=0.34.0",
    
    # IPFS and multiformats
    "pinata>=0.0.1",
    "multiformats>=0.3.1.post4",
    "py-multicodec>=0.2.1",
    "py-multiformats-cid>=0.4.4",
    
    # LangGraph and LangChain dependencies (for specialized agents)
    "langgraph>=0.2.59",
    "langchain-core>=0.3.30",
    "langchain-openai>=0.1.0",
    "langchain-mcp-adapters>=0.1.1",
    "pydantic>=2.0.0",
    
    # Utilities
    "loguru>=0.7.3",
    "ruff>=0.9.7",
]

[project.optional-dependencies]
test = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.1",
    "pytest-asyncio>=0.21.1",
    "responses>=0.24.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
filterwarnings = [
    "ignore::DeprecationWarning",
]
addopts = "--cov=. --cov-report=term-missing"

[dependency-groups]
dev = [
    "pytest-cov>=6.0.0",
]
